Make it explicit calendar types are for events
