+++ Packages/EurofurenceComponents/Sources/EventDetailComponent/Calendar/CalendarEventRepository.swift
@@ -1,5 +1,7 @@
+import EurofurenceModel
+
 public protocol CalendarEventRepository {
     
-    func calendarEvent(for identifier: String) -> CalendarEvent
+    func calendarEvent(for identifier: EventIdentifier) -> CalendarEvent
     
 }
