+++ Packages/EurofurenceComponents/Tests/EventDetailComponentTests/View Model Tests/WhenEventIsPresentInDeviceCalendar_EventDetailViewModelShould.swift
@@ -2,64 +2,12 @@ import EventDetailComponent
 import XCTest
 import XCTEurofurenceModel
 
-class FakeCalendarEventRepository: CalendarEventRepository {
-    
-    private var events = [String: FakeCalendarEvent]()
-    
-    func calendarEvent(for identifier: String) -> CalendarEvent {
-        if let event = events[identifier] {
-            return event
-        } else {
-            let event = FakeCalendarEvent()
-            events[identifier] = event
-            
-            return event
-        }
-    }
-    
-    func fakedEvent(for identifier: String) -> FakeCalendarEvent? {
-        events[identifier]
-    }
-    
-}
-
-class FakeCalendarEvent: CalendarEvent {
-    
-    enum CalendarAction: Equatable {
-        case unset
-        case added
-        case removed
-    }
-    
-    weak var delegate: CalendarEventDelegate?
-    
-    private(set) var lastAction: CalendarAction = .unset
-    func addToCalendar() {
-        lastAction = .added
-        delegate?.calendarEvent(self, presenceDidChange: .present)
-    }
-    
-    func removeFromCalendar() {
-        lastAction = .removed
-        delegate?.calendarEvent(self, presenceDidChange: .absent)
-    }
-    
-    func simulateEventPresent() {
-        delegate?.calendarEvent(self, presenceDidChange: .present)
-    }
-    
-    func simulateEventAbsent() {
-        delegate?.calendarEvent(self, presenceDidChange: .absent)
-    }
-    
-}
-
 class WhenEventIsPresentInDeviceCalendar_EventDetailViewModelShould: XCTestCase {
     
     func testIncludeRemoveFromCalendarAction() throws {
         let event = FakeEvent.random
         let context = EventDetailViewModelFactoryTestBuilder().build(for: event)
-        let calendarEvent = try XCTUnwrap(context.calendarEventRepository.fakedEvent(for: event.identifier.rawValue))
+        let calendarEvent = try XCTUnwrap(context.calendarEventRepository.fakedEvent(for: event.identifier))
         calendarEvent.simulateEventPresent()
         let visitor = context.prepareVisitorForTesting()
         let command = visitor.visited(ofKind: AddEventToCalendarAction.self)
@@ -73,7 +21,7 @@ class WhenEventIsPresentInDeviceCalendar_EventDetailViewModelShould: XCTestCase 
     func testRemoveEventFromCalendarWhenInvokingAction() throws {
         let event = FakeEvent.random
         let context = EventDetailViewModelFactoryTestBuilder().build(for: event)
-        let calendarEvent = try XCTUnwrap(context.calendarEventRepository.fakedEvent(for: event.identifier.rawValue))
+        let calendarEvent = try XCTUnwrap(context.calendarEventRepository.fakedEvent(for: event.identifier))
         calendarEvent.simulateEventPresent()
         let visitor = context.prepareVisitorForTesting()
         let command = visitor.visited(ofKind: AddEventToCalendarAction.self)
@@ -85,7 +33,7 @@ class WhenEventIsPresentInDeviceCalendar_EventDetailViewModelShould: XCTestCase 
     func testShowTheAddToCalendarActionAfterRemovingEventFromCalendar() throws {
         let event = FakeEvent.random
         let context = EventDetailViewModelFactoryTestBuilder().build(for: event)
-        let calendarEvent = try XCTUnwrap(context.calendarEventRepository.fakedEvent(for: event.identifier.rawValue))
+        let calendarEvent = try XCTUnwrap(context.calendarEventRepository.fakedEvent(for: event.identifier))
         calendarEvent.simulateEventPresent()
         let visitor = context.prepareVisitorForTesting()
         let command = visitor.visited(ofKind: AddEventToCalendarAction.self)
